Move Range methods to IntOps
