--- semtypes/src/main/java/io/ballerina/semtype/RangeValue.java
@@ -1,37 +1,0 @@
-/*
- *  Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package io.ballerina.semtype;
-
-/**
- * Wraps the empty range union / intersect.
- * -1 means union/intersect is empty because r1 is before r2, with no overlap
- * 1 means union/intersect is empty because r2 is before r1 with no overlap
- *
- * @since 2.0.0
- */
-public class RangeValue implements RangeHolder {
-    public final int value;
-
-    private RangeValue(int value) {
-        this.value = value;
-    }
-
-    public static RangeValue from(int value) {
-        return new RangeValue(value);
-    }
-}
